refactor:rename from auth_usecase to auth
