+++ lib/features/auth/domain/usecases/auth.dart
