+++ lib/features/auth/presentation/controllers/auth_controller.dart
@@ -1,7 +1,7 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:trip_app_nativeapp/core/view/widgets/loading.dart';
-import 'package:trip_app_nativeapp/features/auth/domain/usecases/auth_usecase.dart';
+import 'package:trip_app_nativeapp/features/auth/domain/usecases/auth.dart';
 
 import '../../../../core/exception/exception_handler.dart';
 import '../../../../core/view/widgets/helpers/scaffold_messenger.dart';
